requirements.txt
click==8.1.3
Flask==2.1.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.20.0
PyYAML==6.0
wakeonlan==2.1.0
Werkzeug==2.1.2
zipp==3.8.1
